Drivetrain (#1)
* drivetrain en telOP niet getest

* fixed drivetrain en deleted all examples

* fixed drivetrain en deleted all examples 2.0

* deleted the speedLowering for the drivetrain

---------
