+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/subsystem/MecanumDrivetrain.java
@@ -4,41 +4,43 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
-import org.firstinspires.ftc.teamcode.lib.kinematics.ChassisSpeeds;
-
 
 /**
- * This file is an example of how to define a tank drivetrain in your code.
+ * This file is an example of how to define a mecanum drivetrain in your code.
  * It is structured in the same way as the ExampleSubsystem
  *
  * Different functions of the drivetrain have been defined, but not yet fully worked out
  */
-public class ExampleTankDrivetrain {
+public class MecanumDrivetrain {
 
     //Declare motor and servo objects
-    private DcMotor rightMotor;
-    private DcMotor leftMotor;
+    private DcMotor rightFront;
+    private DcMotor leftFront;
+    private DcMotor rightBack;
+    private DcMotor leftBack;
 
 
     /**
      * This is the constructor of the subsystem
      * This is the function that will be run when the subsystem is created,
      * which happens at the beginning of an OpMode.
-     * The constructor should have the same name as the class (ExampleTankDrivetrain in this case).
+     * The constructor should have the same name as the class (ExampleMecanumDrivetrain in this case).
      *
      * @param hardwareMap This is the input of the constructor, which will be used
      *                    to link the motors and servos in the code to the motors and servos
      *                    on the actual robot
      */
-    public ExampleTankDrivetrain(HardwareMap hardwareMap){
+    public MecanumDrivetrain(HardwareMap hardwareMap){
         /*
          * These lines of code links the DcMotors to the ports on the control/expansion hub
          * with the corresponding labels (deviceName)
          * This 'labeling' can be done on the Driver Station by clicking on the three dots
          * in the upper right corner and then going to 'Configure Robot'
          */
-        rightMotor = hardwareMap.get(DcMotor.class, "right_front");
-        leftMotor =  hardwareMap.get(DcMotor.class, "left_front");
+        rightFront = hardwareMap.get(DcMotor.class, "motor0");
+        leftFront =  hardwareMap.get(DcMotor.class, "motor1");
+        rightBack =  hardwareMap.get(DcMotor.class, "motor2");
+        leftBack =  hardwareMap.get(DcMotor.class, "motor3");
 
         /*
          * Normally a DC motors runs in the clockwise direction for positive values
@@ -47,14 +49,17 @@ public class ExampleTankDrivetrain {
          * Usually the left side of the drivetrain needs to be reversed,
          * but this should always be checked to be sure
          */
-        leftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+        leftBack.setDirection(DcMotorSimple.Direction.REVERSE);
+        leftFront.setDirection(DcMotorSimple.Direction.REVERSE);
 
         /*
          * Tell the motors to use the integrated encoders
          * This gives a bit more precision while controlling the motors
          */
-        rightMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        leftMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        rightFront.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        leftFront.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        rightBack.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        leftBack.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
     }
 
     /*
@@ -64,27 +69,26 @@ public class ExampleTankDrivetrain {
      */
 
 
-    /**
-     * Function to make the robot drive forward/backward and possibly rotate
-     *
-     * @param forwardSpeed How fast the robot needs to move forward (use negative speeds to go backward)
-     * @param rotationSpeed How fast the robot needs to rotate (positive is clockwise)
-     */
-    public void tankDrive(double forwardSpeed, double rotationSpeed){
 
-        // Calculate the motor speed for left and right based on the forward and rotation speed
-        double rightPower = forwardSpeed - rotationSpeed;
-        double leftPower = forwardSpeed + rotationSpeed;
 
-        rightMotor.setPower(rightPower);
-        leftMotor.setPower(leftPower);
+
+    public void mecanumDrive(double x, double y, double rx){
+        leftFront.setPower((y + x + rx));
+        leftBack.setPower((y - x + rx));
+        rightFront.setPower((y - x - rx));
+        rightBack.setPower((y + x - rx));
     }
-
     /**
      * Stop all motors of the drivetrain
      */
     public void stop(){
-        rightMotor.setPower(0);
-        leftMotor.setPower(0);
+        rightFront.setPower(0);
+        leftFront.setPower(0);
+        rightBack.setPower(0);
+        leftBack.setPower(0);
     }
+    public void setLeftBackSpeed(double speed){leftBack.setPower(speed);}
+    public void setRightBackSpeed(double speed){rightBack.setPower(speed);}
+    public void setLeftFrontSpeed(double speed){leftFront.setPower(speed);}
+    public void setRightFrontSpeed(double speed){rightFront.setPower(speed);}
 }
